fix(scraper): move service_controller out of services/railway/ package
The services/railway/ package's __init__.py eagerly imports SQLAlchemy via
connection.py, which crashes the Railway cron runtime on first import (parent
package gets rolled back from sys.modules, yielding a misleading
"No module named 'services.railway'" error).

Moving the file to services/railway_service_controller.py (flat path) bypasses
the buggy __init__.py entirely. Tests pass unchanged; only the import path
moves.

Caught by deployment crash on first cron run with the new code; CI didn't
catch it because dev-environment deps were complete.
